# openapi/bifrost-api.yaml
openapi: 3.0.3
info:
  title: Bifrost Admin API
  description: Edge router management API for route configuration and analytics
  version: 1.38.0
  contact:
    name: Bifrost Maintainers
    url: https://example.com
servers:
  - url: https://bifrost.example.com
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /api/routes:
    get:
      summary: List routes or get single route
      parameters:
        - $ref: '#/components/parameters/PathQuery'
        - $ref: '#/components/parameters/DomainQuery'
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
        - $ref: '#/components/parameters/RouteTypeQuery'
        - $ref: '#/components/parameters/RouteEnabledQuery'
      responses:
        '400':
          description: >-
            An invalid list query (a limit or offset that is not a whole number
            in range, an unknown type or enabled value, a search over 2,048
            characters): refused, never answered with an unfiltered list
        '200':
          description: >-
            The route (with `path`), or the list. A stored record that cannot be
            read is listed as a minimal row `{ domain, path, invalid: true }`,
            after the readable routes; it matches a search by its path only and
            no type or enabled filter. A search in a one-domain list does not
            match the domain.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoutesResponse'
        '409':
          description: >-
            The route asked for by `path` is stored in a shape that cannot be read
            (`error: ROUTE_RECORD_INVALID`); delete it and create it again
    post:
      summary: Create route
      parameters:
        - $ref: '#/components/parameters/DomainQueryWrite'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteInput'
      responses:
        '201':
          description: Created
        '409':
          description: >-
            A route already exists at the path, or the stored record there cannot be
            read (`error: ROUTE_RECORD_INVALID`; delete it and create it again)
    put:
      summary: Update route
      parameters:
        - $ref: '#/components/parameters/PathQueryRequired'
        - $ref: '#/components/parameters/DomainQueryWrite'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteInput'
      responses:
        '200':
          description: Updated
        '409':
          description: >-
            The stored record cannot be read (`error: ROUTE_RECORD_INVALID`); it is
            never merged with the patch. Delete it and create it again
    delete:
      summary: Delete route
      parameters:
        - name: path
          in: query
          required: true
          description: >-
            Route path (required). With `recover=invalid`, the stored key's own
            path exactly as a listing shows it, never normalised.
          schema:
            type: string
            pattern: ^/.*
        - $ref: '#/components/parameters/DomainQueryWrite'
        - name: recover
          in: query
          required: false
          description: >-
            `invalid`: the exact-key recovery of an unreadable record. Deletes
            the record stored at exactly `{domain}:{path}`, only when it cannot
            be read; a readable route there is refused (409
            `ROUTE_RECORD_READABLE`). Without it the path is normalised, as
            every route write normalises it.
          schema:
            type: string
            enum: [invalid]
      responses:
        '200':
          description: >-
            Deleted, also a route whose stored record cannot be read (its public
            URL is purged, as stored and percent-encoded for a recovery, and its
            audit row names its key with `state: invalid`)
        '404':
          description: Nothing is stored at the key
        '409':
          description: >-
            A recovery aimed at a readable route (`error: ROUTE_RECORD_READABLE`);
            delete it with the ordinary delete
...
  /api/qr/{id}:
    get:
      summary: Get a QR code record
      tags: [QR Codes]
      parameters:
        - $ref: '#/components/parameters/QrIdPath'
        - $ref: '#/components/parameters/DomainQuery'
      responses:
        '200':
          description: The QR code record
        '404':
          $ref: '#/components/responses/QrNotFound'
        '409':
          $ref: '#/components/responses/QrRecordInvalid'
    put:
      summary: Update a QR code (type is immutable)
      tags: [QR Codes]
      parameters:
        - $ref: '#/components/parameters/QrIdPath'
        - $ref: '#/components/parameters/DomainQueryWrite'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: QR code updated
        '404':
          $ref: '#/components/responses/QrNotFound'
        '409':
          $ref: '#/components/responses/QrRecordInvalid'
    delete:
      summary: Delete a QR code (hard delete; the audit log preserves the record)
      tags: [QR Codes]
      parameters:
        - $ref: '#/components/parameters/QrIdPath'
        - $ref: '#/components/parameters/DomainQueryWrite'
      responses:
        '200':
          description: >-
            QR code deleted: `{ success: true, data: { deleted: true, id,
            createdAt } }`. `createdAt` is the deleted record's, naming the
            incarnation removed (absent when the record could not be read);
            a code re-created later with the same id has another
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    required: [deleted, id]
                    properties:
                      deleted:
                        type: boolean
                        enum: [true]
                      id:
                        type: string
                      createdAt:
                        type: number
        '404':
          $ref: '#/components/responses/QrNotFound'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  responses:
    QrNotFound:
      description: >-
        The QR code does not exist: `{ success: false, error: QR_NOT_FOUND,
        message }`.
      content:
        application/json:
          schema:
            type: object
            required: [success, error, message]
            properties:
              success:
                type: boolean
                enum: [false]
              error:
                type: string
                enum: [QR_NOT_FOUND]
              message:
                type: string
    QrRecordInvalid:
      description: >-
        The QR code is stored in a shape that cannot be read: `{ success: false,
        error: QR_RECORD_INVALID, message }`. It exists, so this is never a
        deletion; delete it and create it again.
      content:
        application/json:
          schema:
            type: object
            required: [success, error, message]
            properties:
              success:
                type: boolean
                enum: [false]
              error:
                type: string
                enum: [QR_RECORD_INVALID]
              message:
                type: string
  parameters:
    QrIdPath:
      name: id
      in: path
      required: true
      description: QR code id (lowercase slug or generated hex)
      schema:
        type: string
    PathQuery:
      name: path
      in: query
      description: Route path for single route lookup
      schema:
        type: string
        pattern: ^/.*
    PathQueryRequired:
      name: path
      in: query
      required: true
      description: Route path (required)
      schema:
        type: string
        pattern: ^/.*
    DomainQuery:
      name: domain
      in: query
      description: Target domain
      schema:
        type: string
        enum:
          - example.com
          - links.example.com
          - bifrost.example.com
          - secondary.example.net
          - user1.example.com
          - user2.example.com
          - user3.example.com
          - couple.example.com
          - user5.example.com
    DomainQueryWrite:
      name: domain
      in: query
      description: >-
        Target domain of the write. Required here or in the `X-Domain` header,
        and the two must match when both are sent; a write that names neither,
        or names conflicting domains, is refused with 400 and nothing is
        written. There is no default domain for writes.
      schema:
        type: string
        enum:
          - example.com
          - links.example.com
          - bifrost.example.com
          - secondary.example.net
          - user1.example.com
          - user2.example.com
          - user3.example.com
          - couple.example.com
          - user5.example.com
...
    LimitQuery:
      name: limit
      in: query
      description: Results per page
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    OffsetQuery:
      name: offset
      in: query
      description: Pagination offset
      schema:
        type: integer
        minimum: 0
        default: 0
    RouteTypeQuery:
      name: type
      in: query
      description: Only routes of this type (an unreadable record matches none)
      schema:
        type: string
        enum: [redirect, proxy, r2]
    RouteEnabledQuery:
      name: enabled
      in: query
      description: Only enabled (`true`) or disabled (`false`) routes (an unreadable record matches neither)
      schema:
        type: string
        enum: ['true', 'false']
...
    SearchQuery:
      name: search
      in: query
      description: >-
        Search over path, target, type, status code, bucket and host header,
        ignoring case and separators (spaces, hyphens, underscores, dots,
        slashes), words in any order within one field; the domain matches as
        typed. Matches are ordered by relevance (path matches first), newest
        first on ties. Matching reads the first 200 characters; a longer value
        than 2,048 characters answers 400.
      schema:
        type: string
        maxLength: 2048
...
  schemas:
...
    CreateRouteInput:
      type: object
      required:
        - path
        - type
        - target
      properties:
        path:
          type: string
          minLength: 1
          pattern: ^/.*
        type:
          type: string
          enum: [redirect, proxy, r2]
        target:
          type: string
          minLength: 1
          maxLength: 8192
        statusCode:
          type: integer
          enum: [301, 302, 307, 308]
        preserveQuery:
          type: boolean
          default: true
        preservePath:
          type: boolean
          default: false
        cacheControl:
          type: string
          maxLength: 256
        hostHeader:
          type: string
          maxLength: 253
          description: Override Host header for proxy requests
        forceDownload:
          type: boolean
          default: false
        bucket:
          type: string
          enum: [files, assets, files-user1, files-user2, files-user3, files-user4, files-user5, files-user6]
          default: files
          description: R2 bucket for file serving (R2 only)
        enabled:
          type: boolean
          default: true
    UpdateRouteInput:
      type: object
      properties:
        type:
          type: string
          enum: [redirect, proxy, r2]
        target:
          type: string
          maxLength: 8192
        statusCode:
          type: integer
          enum: [301, 302, 307, 308]
        preserveQuery:
          type: boolean
        preservePath:
          type: boolean
        cacheControl:
          type: string
          maxLength: 256
        hostHeader:
          type: string
          maxLength: 253
          description: Override Host header for proxy requests
        forceDownload:
          type: boolean
        bucket:
          type: string
          enum: [files, assets, files-user1, files-user2, files-user3, files-user4, files-user5, files-user6]
          description: R2 bucket for file serving (R2 only)
        enabled:
          type: boolean
...
    RoutesResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        error:
          type: string
...